Renamed domain package into entity. Added interfaces: CityDAOImpl, CountryDAOImpl Added Logger and log4j2.xml configuration, RedisClientProvider, RedisDataMapper, MainAppService; 	SQLDb.dump in resources folder. Refactored: MainApp.
