+++ pom.xml
@@ -17,6 +17,9 @@
         <p6spy.version>3.9.1</p6spy.version>
         <lettuce-core.version>6.2.2.RELEASE</lettuce-core.version>
         <jackson-databind.version>2.14.0</jackson-databind.version>
+        <log4j-api.version>2.17.2</log4j-api.version>
+        <log4j-core.version>2.17.2</log4j-core.version>
+        <log4j-slf4j.version>2.17.2</log4j-slf4j.version>
     </properties>
 
     <dependencies>
@@ -48,6 +51,24 @@
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
             <version>${jackson-databind.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.logging.log4j</groupId>
+            <artifactId>log4j-api</artifactId>
+            <version>${log4j-api.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.logging.log4j</groupId>
+            <artifactId>log4j-core</artifactId>
+            <version>${log4j-core.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.logging.log4j</groupId>
+            <artifactId>log4j-slf4j-impl</artifactId>
+            <version>${log4j-slf4j.version}</version>
         </dependency>
 
     </dependencies>
