+++ src/main/java/ru/javarush/module4/projecthibernatefinal/dao/CountryDAO.java
@@ -2,11 +2,11 @@ package ru.javarush.module4.projecthibernatefinal.dao;
 
 import org.hibernate.SessionFactory;
 import org.hibernate.query.Query;
-import ru.javarush.module4.projecthibernatefinal.domain.Country;
+import ru.javarush.module4.projecthibernatefinal.entity.Country;
 
 import java.util.List;
 
-public class CountryDAO {
+public class CountryDAO implements CountryDAOImpl {
     private final SessionFactory sessionFactory;
 
     public CountryDAO(SessionFactory sessionFactory) {
